pyproject.toml
[tool.poetry]
name = "langchain-chatbot-api"
version = "0.1.0"
description = ""
authors = ["znu <znu4847@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
pyjwt = "^2.9.0"
django-environ = "^0.11.2"
dj-database-url = "^2.3.0"
psycopg2-binary = "^2.9.10"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
gunicorn = "^23.0.0"
uvicorn = "^0.32.0"
channels = "^4.2.0"
daphne = "^4.1.2"
langchain = "^0.3.14"
langchain-core = "^0.3.29"
langchain-openai = "^0.2.14"
langchain-community = "^0.3.14"
django-cors-headers = "^4.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
